stub out wp api client
