+++ .env.example
@@ -3,6 +3,7 @@ VUE_APP_API_BASE_URL=http://localhost:8080/
 VUE_APP_LINK_BASE_URL=https://test.americanwhitewater.org
 VUE_APP_MAPBOX_ACCESS_TOKEN=
 VUE_APP_NWI_TILE_SERVER=http://localhost:8080/tiles/{z}/{x}/{y}.mvt/
+VUE_APP_WP_API_URL=
 
 VUE_APP_APOLLO_DEV_HELPER=
 VUE_APP_CLIENT_ID=
