+++ src/app/environment/index.js
@@ -36,7 +36,7 @@ const clientId = VUE_APP_CLIENT_ID
 const clientSecret = VUE_APP_CLIENT_SECRET
 const mapboxAccessToken = VUE_APP_MAPBOX_ACCESS_TOKEN
 const nwiTileServer = VUE_APP_NWI_TILE_SERVER
-const cmsBaseUrl = VUE_APP_WP_API_URL
+const wpApiUrl = VUE_APP_WP_API_URL
 const laravelDeploy = Boolean(VUE_APP_LARAVEL_DEPLOY)
 const baseUrl = VUE_APP_LINK_BASE_URL
 
@@ -48,7 +48,7 @@ export {
   assetBaseUrl,
   mapboxAccessToken,
   nwiTileServer,
-  cmsBaseUrl,
+  wpApiUrl,
   clientId,
   clientSecret,
   laravelDeploy,
