+++ src/app/services/canCreateReach.js
@@ -1,10 +1,10 @@
-import http from "@/app/http";
+import { laravelClient } from "@/app/http";
 
 export async function canCreateReach({ plat, plon }) {
   /**
    * below the info requires an object literal, as variable interpolation is not part of the type "ArbitraryJSON"
    */
-  return http
+  return laravelClient
     .post("graphql", {
       query: `
       query{
