+++ src/app/services/createReach.js
@@ -1,8 +1,8 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function createReach(data) {
   
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
         mutation ($reach: ReachInput!) {
           reachUpdate(reach: $reach) {
