+++ src/app/services/deleteReach.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function deleteReach(id) {
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
             mutation ($id:ID!) {
               reachDelete(id: $id) {
