+++ src/app/services/getAffiliate.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getAffiliate(id) {
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
       query {
         affiliate(id: !ID) {
