+++ src/app/services/getArticles.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getArticles() {
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
         query {
           articles(first: 25, orderBy: {field: POSTED_DATE, order: ASC}, page: 1) {
