+++ src/app/services/getDocuments.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getDocuments() {
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
       query {
         __typename
