+++ src/app/services/getGage.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getGage(id) {
-  return http.post('graphql', {
+  return laravelClient.post('graphql', {
     query: `
       query ($gauge_id: ID!) {
         gauge(id: $gauge_id) {
