+++ src/app/services/getGageReadings.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getGageReadings(data) {
   const url = `api/gauge/${data.gauge_id}/flows/${data.metric_id}?from=${data.timeStart}&to=${data.timeEnd}&resolution=${data.resolution}`
   
-  return http.get(url).then(res => res.data)
+  return laravelClient.get(url).then(res => res.data)
 }
