+++ src/app/services/getReachEvents.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getReachEvents(id) {
-  return http
+  return laravelClient
     .post('graphql', {
       query: `
         query {
