+++ src/app/services/getReachGages.js
@@ -1,7 +1,7 @@
-import http from "@/app/http";
+import { laravelClient } from "@/app/http";
 
 export async function getReachGages(id) {
-  return http
+  return laravelClient
     .post("/graphql", {
       query: `{
         getGaugeInformationForReachID(id: ${id}) {
