+++ src/app/services/getStateList.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function getStateList() {
-  return http.post('/graphql', {
+  return laravelClient.post('/graphql', {
     query: `
             query {
                   states(first:150,aw_only:true){data{gmi,name,shortkey,type,aw_region,num_rivers,num_gauges}}
