+++ src/app/services/searchReachMap.js
@@ -1,7 +1,7 @@
-import http from "@/app/http"
+import { laravelClient } from "@/app/http"
 
 export async function searchReachMap(data) {
-  return http.post('graphql/', {
+  return laravelClient.post('graphql/', {
     query: `
       {
         reachmap(first: 50, match: "${data}") {
