+++ src/app/store/modules/news-page.js
@@ -1,6 +1,6 @@
 import actions from '@/app/store/actions'
 import mutations from '@/app/store/mutations'
-import http from '@/app/http'
+import { laravelClient } from '@/app/http'
 
 /**
  * this setup is for headless wordpress. 
@@ -19,16 +19,16 @@ export default {
   actions: {
     ...actions,
     getNewsArticles: () => {
-      return http.get('/v2/posts').then(res => res.data)
+      return laravelClient.get('/v2/posts').then(res => res.data)
     },
     searchArticles: () => {
-      return http.get('/v2/posts?content=').then(res => res.data)
+      return laravelClient.get('/v2/posts?content=').then(res => res.data)
     },
     getFeaturedMedia: () => {
-      return http.get('/v2/media/').then(res => res.data)
+      return laravelClient.get('/v2/media/').then(res => res.data)
     },
     getFrontPageNews: () => {
-      return http.get('frontpage').then(res => res.data)
+      return laravelClient.get('frontpage').then(res => res.data)
     }
   }
 }
