+++ src/app/store/modules/user.js
@@ -1,7 +1,7 @@
 import actions from '@/app/store/actions'
 import mutations from '@/app/store/mutations'
-import http from "@/app/http"
-import {getUser} from "@/app/services"
+import { laravelClient } from "@/app/http"
+import { getUser } from "@/app/services"
 
 export default {
   namespaced: true,
@@ -26,12 +26,12 @@ export default {
     },
     userForgot: async (data) => {
 
-      return http.post('graphql', data).then(res => res.data)
+      return laravelClient.post('graphql', data).then(res => res.data)
     },
     userLogin: data => {
       
 
-      return http.post('graphql', {
+      return laravelClient.post('graphql', {
         query: `
         mutation {
             login(input: ${data}) {
@@ -45,7 +45,7 @@ export default {
     },
     userRegister: data => {
 
-      return http.post('graphql', data).then(res => res.data)
+      return laravelClient.post('graphql', data).then(res => res.data)
     },
     logout: (context) => {
       context.commit('DATA_RESET');
